rename database-util and adjust code-style on app.py
